backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annoy>=1.17.3",
    "fastapi[standard]>=0.115.12",
    "geopy>=2.4.1",
    "meteostat>=1.6.8",
    "numpy>=2.2.4",
    "pinecone>=6.0.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pywikibot>=10.0.0",
    "rasterio>=1.4.3",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
]
